Add solver with tests (used with permission)
This is a port of the Nonogram solver by fedimser and used with permission.
https://fedimser.github.io/nonogram
https://github.com/fedimser/fedimser.github.io/blob/master/js/Nonogram.js

We need the solver to be used in conjunction with the random generator of puzzles so that way we can confirm if a puzzle is uniquely solveable. If not, then we generate a new one, and repeat.