Create toKey method that joins all values as string
Instead of doing the manual string concatenation, we can use a method that takes rest parameters so we can provide as many values as we want to generate a string for React's key attribute. In one case, a value will be an array which is why first we make sure to get a toString representation of everything before calling join.