Simplify selection code for pointer events
We can now remove a ton of code as our pointer event setup is much simpler. Also simplifies the logic for determining the cell state (e.g. filled, crossed).

This is the final result of replacing the mouse/touch events with pointer events and maintaining the same behavior as before. The one difference is for all pointeres, when tapping the same cell multiple times, it toggles between the 3 states: filled, crossed and empty. This way it plays well on mobile using one hand. No other buttons necessary to toggle cell state selection for example.