Remove condition for answer result if crossed
No need to do this now since we create the answer matrix with default value of 0 for all positions. Otherwise to answer the puzzle one would have to set the state for every cell to filled or crossed but that isn't necessary. What matters if the cells that are filled matched the solution.