Add title attribute on nonogram grid cells
This was an attempt to use testing-library to simulate user events but that causes failures since some APIs are not implemented in jsdom which makes sense. We'll scrap that for now. Once we split out the components we can try some tests with mocks/stubs for fun.

For now we'll have the title attribute in as its helpful to know what row/column it is.