Use named exports
Using default exports has its place but for our own code we'll prefer to use named exports for consistency. This way its explicit and imported with the same name as defined. The file importing a module can now be explicit if it wants to rename the export to something else. So we don't lose anything but is more explicit/predictable now.

We'll update the solvable export separately.