Fix error on pointer moving past boundaries on mobile
Despite having a pointerLeave handler, this can still fail on mobile when pointing down and moving across the grid and out of the pointer area. Since we already handle the null case in our event handlers, we just need to update this method to only process the target if we found one depending on the pointer type, otherwise nothing is returned which is what we want.