Extract to Puzzle component with fixed data swap on complete
This extracts from the App, a Puzzle component that has everything it needs to render and a puzzle with pointer events for behavior. The App component now has the data and an on complete handler. Since we already know when a puzzle is answered correctly, we want to test the next step which is rendering a brand new puzzle.

For now, we'll use a fixed second puzzle to test which is working. Since we duplicated the puzzle to be a 10x5, what happened is the first half of the puzzle on completion remained unchanged. Since the component key didn't change (the data given as the key was the same) React didn't update the elements which is right but not what we wanted. To fix this, at the parent level, the Puzzle component which renders the row hints, column hints and cells, we have to specify the data as part of the key. This way when it changes, React will re-render everything and show a brand new puzzle even if some of the data remains the same.

In the future, we can change this to use a puzzle ID (e.g. a hash of the data array) for the component key argument.