Support grids 5x10 and 10x5 with default as 5x5
If no grid-NxN class given/found then a 5x5 grid is assumed. This implements 2 classes, grid-5x10 and grid-10x5 that just changes the number of rows and columns on the nonogram grid. Since we are using auto placement on the row and column hint grids, they will automatically line up as expected without having to change the grid template columns/rows for them. This is great as its all derived from the puzzle grid (the main data set) as it should be.