Update column hint test case
Since we changed the default puzzle in our app, there are multiple hints of 5 so this throws an error if multiple matches are found. This isn't a helpful test much if we change the puzzle but just to have something here for now so we can expand and learn how to with Jest/client side.