Major update: component class, endless seeded random generation
This updates the Puzzle functional component to a class based component. I believe I started this because it was clearer when adding binded event handlers, otherwise this didn't behave as expected.

This also added endless random generation of new puzzles when completing one. I realized it was easy to add in once I had the random generation up and solver as well. It's bsed off a seed that is hard coded for now for testing purposes.