Fix pointerLeave bug on completing puzzle
If the puzzle is answered correctly with the last pointer down/move action but you happen to pointer up outside the grid (triggers pointer leave) the puzzle remains in a completed state without resetting. You'd have to toggle any cell back to what it was to trigger the completedAction. The reason for this is that the pointerLeaveHandler would remove both the pointer move and up handlers. It should just remove the pointer move handler.