Isolate puzzle gen and build hints as array of objects
For hints, we want to change the hint number that part of the puzzle has been solved/met but to do that we need to calculcate not just the hint number, but the start and length of where the hint applies in the puzzle line. With this, in the HintGroup component when building out the hint numbers, we check if the hint is solved by slicing the line of data specific to that hint. If everything is 1 (selected cell) then its solved. There is a special case for when the hint is 0 and thus the only hint. The solution check is just modified slightly to account for when total is 0 and to instead check for 0 as data value (empty/crossed cell).

The other major change here is to isolate the puzzle generation to a parent component. While we could have it all together I think this may be better because the puzzle doesn't change once its generated, so giving that data as props to the Board is ideal. The Board component then is just responsible for tracking the users' solution to the puzzle which is the only state that changes now. When a new puzzle is generated, by including a key attribute, React will recreate the component tree. This is the simplest solution to ensure a fresh start for now, otherwise we'll have to recreate the Board answer state when the data changes which I don't have a good handle on yet.